+++ src/features/channels/stories/ChannelsEdit.stories.tsx
@@ -1,7 +1,6 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 import React from "react";
-
-import { ChannelsEdit } from "./ChannelsEdit";
+import { ChannelsEdit } from "../components";
 
 const channelsList = ["conda-store", "default", "conda forge"];
