+++ src/features/channels/stories/ChannelsList.stories.tsx
@@ -1,7 +1,6 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 import React from "react";
-
-import { ChannelsList } from "./ChannelsList";
+import { ChannelsList } from "../components";
 
 const channelsList = ["conda-store", "default", "conda forge"];
